+++ apps/alpha-test-app/CHANGELOG.md
@@ -1,5 +1,19 @@
 # alpha-test-app
 
+## 0.1.0-alpha.116
+
+### Patch Changes
+
+- Updated dependencies [697ffcd]
+  - sit-onyx@1.0.0-alpha.119
+
+## 0.1.0-alpha.115
+
+### Patch Changes
+
+- Updated dependencies [2d2ada6]
+  - sit-onyx@1.0.0-alpha.118
+
 ## 0.1.0-alpha.114
 
 ### Patch Changes