+++ apps/playground/CHANGELOG.md
@@ -1,5 +1,19 @@
 # playground
 
+## 0.0.1-alpha.58
+
+### Patch Changes
+
+- Updated dependencies [697ffcd]
+  - sit-onyx@1.0.0-alpha.119
+
+## 0.0.1-alpha.57
+
+### Patch Changes
+
+- Updated dependencies [2d2ada6]
+  - sit-onyx@1.0.0-alpha.118
+
 ## 0.0.1-alpha.56
 
 ### Patch Changes