+++ packages/chartjs-plugin/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sit-onyx/chartjs-plugin
 
+## 1.0.0-alpha.75
+
+### Patch Changes
+
+- Updated dependencies [697ffcd]
+  - sit-onyx@1.0.0-alpha.119
+
+## 1.0.0-alpha.74
+
+### Patch Changes
+
+- Updated dependencies [2d2ada6]
+  - sit-onyx@1.0.0-alpha.118
+
 ## 1.0.0-alpha.73
 
 ### Patch Changes