+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,17 @@
 # sit-onyx
 
+## 1.0.0-alpha.119
+
+### Patch Changes
+
+- 697ffcd: Removed gap between label and external link icon in the nav item component
+
+## 1.0.0-alpha.118
+
+### Minor Changes
+
+- 2d2ada6: feat(OnyxTable): support vertical/horizontal scrolling
+
 ## 1.0.0-alpha.117
 
 ### Patch Changes